+++ SharePointFramework/ProjectWebParts/src/components/ProjectPhases/ProjectPhase/ProjectPhase.tsx
@@ -3,8 +3,6 @@ import { IProjectPhaseProps } from './types'
 import { useProjectPhase } from './useProjectPhase'
 import { Popover, PopoverSurface, PopoverTrigger } from '@fluentui/react-components'
 import { ProjectPhasePopover } from './ProjectPhasePopover'
-import pSBC from 'shade-blend-color'
-import { hslToHex, hexToHsl } from 'colors-convert'
 
 export const ProjectPhase: FC<IProjectPhaseProps> = (props) => {
   const {
@@ -14,28 +12,10 @@ export const ProjectPhase: FC<IProjectPhaseProps> = (props) => {
     className,
     subTextProps,
     context,
-    phasesLength
+    phasesLength,
+    altPadding,
+    customPhaseColor
   } = useProjectPhase(props)
-
-  const altPadding = {
-    paddingRight: '.125em',
-    paddingLeft: '.125em'
-  }
-
-  const phaseColor = '#89ce87'
-
-  const changeSaturationAndDarken = (hex: string, saturation: number, darken: number) => {
-    const blendedHex = pSBC(darken, hex)
-    let hslColor = hexToHsl(blendedHex)
-    hslColor.s = saturation
-    return hslToHex(hslColor)
-  }
-
-  const customPhaseColor = {
-    '--phase-color': phaseColor,
-    '--phase-current-color': changeSaturationAndDarken(phaseColor, 40, -0.25),
-    '--phase-current-hover-color': changeSaturationAndDarken(phaseColor, 42, -0.45)
-  }
 
   return (
     <>
